WIP/TEST: Playtest
Note: f2c (version 12.02.01) fails to convert some fortran files.

Background:
blas and lapack are fortran libraries.
blas can be optimized for architectures:
intel, goto, openblas, etc.

lapacke is a c interface (thin wrapper) of lapack. (it can use any
optimized library)
cblas is a c interface (thin wrapper) to blas. (it can use any optimized
library, or it is provided along the optimized versions)

lapacke and cblas still require the fortran libraries.

clapack is a f2c version of lapack (and blas), and this does not require
fortran. Of course, it is not optimized. This project is abandoned.